// src/shared/types.ts
export interface Funcionario {
  id: string;
  nome: string;
  emailCorporativo: string;
  setorId: string;
  cargoId: string;
  usuarioId?: string; // Link com Usuário do Sistema (Login)
  ativo: boolean;
  // Metadata
  createdAt: string;
  updatedAt?: string;
}